+++ RSA/EuclidsExtendedAlgorithm.h
@@ -1,1 +1,1 @@
-long long int miniEuclidsExtendedAlgorithm(long long int e, long long int phiN);
+long long int miniEuclidsExtendedAlgorithm(int e, int phiN);
